Database II

+++ controllers/posts.js
@@ -15,15 +15,17 @@ const getAllPosts = async (req, res) => {
 }
 
 const getPostById = async (req, res) => {
-    const postID = req.params.id
+    const postId = parseInt(req.params.id)
 
     try {
-        const selectedPost = await db.Post.findByPk(postID)
+        const selectedPost = await db.Post.findByPk(postId)
         const author = await selectedPost.getUser()
+        const community = await selectedPost.getCommunity()
 
         const response = {
-            ...selectedPost.dataValues,
+            ...selectedPost.toJSON(),
             author,
+            community,
         }
 
         if (selectedPost === null) {
@@ -40,8 +42,7 @@ const getPostById = async (req, res) => {
 }
 
 const createPost = async (req, res) => {
-    const { title, body } = req.body
-    const userId = req.params.id
+    const userId = parseInt(req.params.id)
 
     try {
         const user = await db.User.findByPk(userId)
@@ -50,10 +51,7 @@ const createPost = async (req, res) => {
             throw new Error('User not found')
         }
 
-        const newPost = {
-            title,
-            body,
-        }
+        const newPost = req.body
 
         const createdPost = await user.createPost(newPost)
 
@@ -67,10 +65,19 @@ const createPost = async (req, res) => {
 }
 
 const updatePost = async (req, res) => {
-    const body = req.body
-    const postId = req.params.id
+    const body = {
+        ...req.body,
+        updatedAt: new Date(),
+    }
+    const postId = parseInt(req.params.id)
 
     try {
+        const post = await db.Post.findByPk(postId)
+
+        if (!post) {
+            throw new Error('Post not found')
+        }
+
         await db.Post.update(body, {
             where: {
                 id: postId,
@@ -88,14 +95,21 @@ const updatePost = async (req, res) => {
 }
 
 const deletePost = async (req, res) => {
-    const postId = req.params.id
+    const postId = parseInt(req.params.id)
 
     try {
+        const post = await db.Post.findByPk(postId)
+
+        if (!post) {
+            throw new Error('Post not found')
+        }
+
         await db.Post.destroy({
             where: {
                 id: postId,
             },
         })
+
         res.status(202).send('Post deleted successfully')
     } catch (e) {
         console.error(e)
